Fix unresponsive slider when forward button tapped on the first occasion. See details for desc.
More details in the original feedback:

"The +15 doesn’t change the time when you click it the first time. For example, the minute was 21. I clicked +15, the scrollbar moved, but the time did not change to the 30th minute. I clicked it again, and it jumped to the 45th minute. Going backwards (-15) works on first click. It’s just the initial +15 click that doesn’t change the time."